backend/requirements.txt
fastapi
uvicorn
pydantic
GitPython
langchain
langchain-community
chromadb
sentence-transformers